chore(qa): format, lint, and mypy green; add health checks and guards; task tracking; update docs and CI
- Fix RUF006 by tracking/naming asyncio tasks in orchestrator and claude init
- Resolve mypy dict typing in style system; scoped mypy config for API
- Replace try/except/pass with logging; align with suppress policy
- Annotate ClassVar mutables in config; minor ruff fixes across tools
- Add async task policy doc; diagnostics and QA plans
- Update CI: ratchet, suppress guard; regen status report (all green)

+++ dashboard/tools/queue_select.py
@@ -144,7 +144,7 @@ def select_now_queue(
         model.Add(sum(x[c.id] for c in cand if c.status == "doing") <= int(wip_cap))
 
     # Objective: maximize Σ(score*100)
-    model.Maximize(sum(int(round(c.score * 100)) * x[c.id] for c in cand))
+    model.Maximize(sum(round(c.score * 100) * x[c.id] for c in cand))
 
     solver = cp_model.CpSolver()
     solver.parameters.max_time_in_seconds = 0.05
